refactor: Restructure category API routes into modular components
- Removed the existing category route implementation and replaced it with a modular structure, separating admin-only and user-accessible routes.
- Introduced a new `admin_only` module for admin-specific operations, including create, update, and delete endpoints.
- Created an `all_users` module to provide paginated category listings for authenticated users.
- Updated the main category router to include these new modules, enhancing code organization and maintainability.
